% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesfm-package.R
\docType{package}
\name{timesfm-package}
\alias{timesfm-package}
\title{timesfm: R Interface to 'TimesFM', a Foundation Model for Time Series Forecasting}
\description{
A thin wrapper around the 'TimesFM' 'Python' package from Google Research (\url{https://github.com/google-research/timesfm}). Provides zero-shot probabilistic forecasting of univariate time series from a pretrained foundation model, with no training or tuning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattyoreilly/timesfm}
  \item Report bugs at \url{https://github.com/mattyoreilly/timesfm/issues}
}

}
\author{
\strong{Maintainer}: Matt O'Reilly \email{fermoymatt@gmail.com}

Authors:
\itemize{
  \item Matt O'Reilly \email{fermoymatt@gmail.com}
}

}
\keyword{internal}
